requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
gunicorn
flask[async]
google.genai
